支持高精度浮点数 允许数组切片具有不同的维度 添加了对索引数量少于维度数量的处理

+++ src/frontend/ast.rs
@@ -211,7 +211,7 @@ pub enum Expression {
 #[derive(Debug, Clone)]
 pub enum Literal {
     IntegerLiteral(i32),
-    FloatLiteral(f32),
+    FloatLiteral(f64), // 使用f64保持高精度
     StringLiteral(String),
     BooleanLiteral(bool),
 }
